Spawn multiple workers
